refactor: `mediaQuery` 폴더를 `utils` -> `constants` 로 변경
