+++ frontend/src/GlobalStyles.js
@@ -1,6 +1,6 @@
 import { css, Global } from '@emotion/react';
 import COLOR from './constants/color';
-import mediaQuery from './utils/mediaQuery';
+import mediaQuery from './constants/mediaQuery';
 
 const GlobalStyles = () => (
   <Global
