+++ frontend/src/components/Items/LevellogItem.styles.ts
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
+import mediaQuery from '../../constants/mediaQuery';
 import { COLOR } from '../../enumerations/color';
-import mediaQuery from '../../utils/mediaQuery';
 
 export const CardStyle = css`
   transition: transform 0.2s ease;
